+++ BIOS_bootloader/src/boot2.asm
@@ -145,7 +145,7 @@ main32:
   mov ds, ax
   mov ss, ax
   mov es, ax
-  mov esp, 0x9000               # stack start at 9000h
+  mov esp, 0x1000               # stack start below the loaded kernels
   # Store kernel parameters
   mov eax, 0x42524F53           # Bootloader Magic value
   mov ebx, SYS_INFO_SEG         # System Info struct address
