+++ kernel/src/arch/x86/APIC.c
@@ -1,0 +1,1 @@
+#include <arch/x86/APIC.h>
