+++ kernel/src/arch/x86/APIC.h
@@ -1,0 +1,1 @@
+#pragma once
